+++ src/onegov.ballot/onegov/ballot/models/election/election_compound.py
@@ -6,7 +6,6 @@ from onegov.ballot.models.election.election_result import ElectionResult
 from onegov.ballot.models.election.list import List
 from onegov.ballot.models.election.list_connection import ListConnection
 from onegov.ballot.models.election.list_result import ListResult
-from onegov.ballot.models.election.mixins import PartyResultsExportMixin
 from onegov.ballot.models.election.panachage_result import PanachageResult
 from onegov.ballot.models.election.party_result import PartyResult
 from onegov.ballot.models.mixins import DomainOfInfluenceMixin
@@ -31,8 +30,7 @@ from sqlalchemy.orm import relationship
 
 class ElectionCompound(
     Base, ContentMixin, TimestampMixin,
-    DomainOfInfluenceMixin, TitleTranslationsMixin,
-    PartyResultsExportMixin
+    DomainOfInfluenceMixin, TitleTranslationsMixin
 ):
 
     __tablename__ = 'election_compounds'
@@ -292,8 +290,8 @@ class ElectionCompound(
                 )
         return rows
 
-    def export_panachage(self):
-        """ Returns all panachage results as list with dicts.
+    def export_parties(self):
+        """ Returns all party results with the panachage as list with dicts.
 
         This is meant as a base for json/csv/excel exports. The result is
         therefore a flat list of dictionaries with repeating values to avoid
@@ -304,18 +302,35 @@ class ElectionCompound(
 
         """
 
-        results = self.panachage_results.order_by(
-            PanachageResult.source,
-            PanachageResult.target,
-            PanachageResult.votes
-        )
+        results = {}
+        parties = set()
+        for result in self.party_results:
+            year = results.setdefault(result.year, {})
+            year[result.name] = {
+                'total_votes': result.total_votes,
+                'color': result.color,
+                'mandates': result.number_of_mandates,
+                'votes': result.votes
+            }
+            parties |= set([result.name])
+        for result in self.panachage_results:
+            year = results.setdefault(self.date.year, {})
+            target = year.setdefault(result.target, {})
+            target[result.source] = result.votes
+            parties |= set([result.source, result.target])
+        parties = sorted(parties)
 
         rows = []
-        for result in results:
-            row = OrderedDict()
-            row['source'] = result.source
-            row['target'] = result.target
-            row['votes'] = result.votes
-            rows.append(row)
+        for year in sorted(results.keys(), reverse=True):
+            for party in sorted(results[year].keys()):
+                row = OrderedDict()
+                row['year'] = year
+                row['name'] = party
+                for column in ('total_votes', 'color', 'mandates', 'votes'):
+                    row[column] = results[year][party].get(column, '')
+                for source in parties:
+                    column = 'panachage_votes_from_{}'.format(source.lower())
+                    row[column] = results[year][party].get(source, '')
+                rows.append(row)
 
         return rows