+++ src/onegov.ballot/onegov/ballot/models/election/mixins.py
@@ -1,5 +1,3 @@
-from collections import OrderedDict
-from onegov.ballot.models.election.party_result import PartyResult
 from sqlalchemy.ext.hybrid import hybrid_property
 
 
@@ -28,38 +26,3 @@ class DerivedAttributesMixin(object):
             return 0
 
         return self.received_ballots / self.eligible_voters * 100
-
-
-class PartyResultsExportMixin(object):
-
-    """ Adds a function to export the party results. """
-
-    def export_parties(self):
-        """ Returns all party results as list with dicts.
-
-        This is meant as a base for json/csv/excel exports. The result is
-        therefore a flat list of dictionaries with repeating values to avoid
-        the nesting of values. Each record in the resulting list is a single
-        candidate result for each political entity. Party results are not
-        included in the export (since they are not really connected with the
-        lists).
-
-        """
-
-        results = self.party_results.order_by(
-            PartyResult.year.desc(),
-            PartyResult.name
-        )
-
-        rows = []
-        for result in results:
-            row = OrderedDict()
-            row['year'] = result.year
-            row['total_votes'] = result.total_votes
-            row['name'] = result.name
-            row['color'] = result.color
-            row['mandates'] = result.number_of_mandates
-            row['votes'] = result.votes
-            rows.append(row)
-
-        return rows